requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
google-genai
apify_client